requirements.txt
# ── Kafka / Streaming ─────────────────────────────────────────────────────────
confluent-kafka==2.4.0
confluent-kafka[avro]==2.4.0
apache-flink==1.19.0

# ── Spark / Data Processing ───────────────────────────────────────────────────
pyspark==3.5.1
pyarrow==16.0.0

# ── Storage / AWS ─────────────────────────────────────────────────────────────
boto3==1.34.0
s3fs==2024.3.1

# ── Data Manipulation ─────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── dbt ───────────────────────────────────────────────────────────────────────
dbt-core==1.8.0
dbt-trino==1.8.0
dbt-utils==1.2.0

# ── Data Quality ──────────────────────────────────────────────────────────────
great-expectations==0.18.19
soda-core-trino==3.3.0

# ── Airflow ───────────────────────────────────────────────────────────────────
apache-airflow==2.9.1
apache-airflow-providers-apache-spark==4.9.0
apache-airflow-providers-apache-kafka==1.3.1

# ── Trino / Query Engine ──────────────────────────────────────────────────────
trino==0.328.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.0
tenacity==8.3.0
structlog==24.1.0
rich==13.7.1
